Added support for ARM and Open Folder Fixes #51 Fixes #44 Fixes #35

+++ src/CompressionResult.cs
@@ -43,7 +43,7 @@ namespace MadsKristensen.ImageOptimizer
         {
             get
             {
-                return Math.Round(100 - (double)ResultFileSize / (double)OriginalFileSize * 100, 1);
+                return Math.Round(100 - ResultFileSize / (double)OriginalFileSize * 100, 1);
             }
         }
